README
This is my first large project on Android and the first one I built with the intention of publishing to the Google Play Store. I spent roughly 4 months developing it and continue to make small improvements and adjustments that I regularly release as updates on the Play Store. I had 3 missions with this project. The first was to create a fully functional and feature rich app that could demonstrate what I've learned while allowing me to develop and practice new skills. The second was to publish my first app and have something tangible that I could show to people to show what I'm capable of. The third was the fact that I had grown tired of bill management apps being bloated with advertising, slow load times, and redirects which made it more difficult than it needed to be to view which bills were due. The idea with this app was to create a streamlined user experience that is aesthetically pleasing and full of detail, without displaying more information than is needed, and instead showing abbreviated data on a user's billers and then allowing them to explore further if needed. This will allow them to manage their bills as quickly as possible and then get back to doing anything other than paying bills! I'm proud of this project and will continue to make adjustments when needed or when I'm not working on other projects. Stay tuned